Multi-Agent-Stock-Analyist/requirements.txt
protobuf
python-dotenv
google-generativeai
google-adk
gunicorn
flask